+++ src/interfaces/AtmPressureEJBRemote.java
@@ -10,7 +10,7 @@ import Sensors.Sensor;
 @Remote
 public interface AtmPressureEJBRemote {
 	List<AtmPressure> findAtmPressure();
-	AtmPressure findAtmPressureById(int id);
+	AtmPressure findAtmPressureById(long id);
 	AtmPressure createAtmPressure(AtmPressure atmPressure);
 	void deleteSensor(AtmPressure atmPressure);
 	AtmPressure updateAtmPressure(AtmPressure atmPressure);
