+++ src/interfaces/UsersEJBRemote.java
@@ -9,7 +9,7 @@ import kernel.Users;
 @Remote
 public interface UsersEJBRemote {
 	List<Users> findUsers();
-	Users findUsersById(int id);
+	Users findUsersById(long id);
 	Users findUsersByDocument(String document);
 	Users createUsers(Users user);
 	void deleteUsers(Users user);
